C-DAT-900-ABJ-2-1-ecp-13/requirements.txt

rasa==3.6.21
rasa-sdk==3.6.2
openai>=1.0.0
python-dotenv
numpy
websockets==10.4
requests